% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill.R
\name{jhbillupdatemain}
\alias{jhbillupdatemain}
\title{同步票据等级}
\usage{
jhbillupdatemain(date, token)
}
\arguments{
\item{date}{日期}

\item{token}{口令}
}
\value{
返回值
}
\description{
同步票据等级
}
\examples{
jhbillupdatemain()
}
